Убрал дату рождения из регистрации
