+++ src/pages/SignupPage/lib/signupValidation.ts
@@ -1,7 +1,4 @@
-import { validateEmail } from 'shared/validation';
-import { validateBirthday } from 'shared/validation';
-import { validatePassword } from 'shared/validation';
-import { validateUsername } from 'shared/validation';
+import { validateEmail, validatePassword, validateUsername } from 'shared/validation';
 import type { ISignupFormData } from '../model/types';
 
 
@@ -17,14 +14,12 @@ import type { ISignupFormData } from '../model/types';
 export const validateSignup = (
     {   email,
         username,
-        birthday,
         password,
         passwordConfirm
     }: ISignupFormData): boolean => {
     const result = {
         email: validateEmail(email).isValid,
         username: validateUsername(username).isValid,
-        birthday: validateBirthday(birthday).isValid,
         password: validatePassword(password).isValid,
         passwordConfirm: password === passwordConfirm && password !== ''
     };
