+++ src/pages/SignupPage/model/types.ts
@@ -1,7 +1,6 @@
 export interface ISignupFormData {
     email: string;
     username: string;
-    birthday: string;
     password: string;
     passwordConfirm?: string;
 }
